Reorganize file transfer related code
