+++ Core/Barcelona/Attachments/CBFileTransfer.swift
