+++ Core/Barcelona/Attachments/IMFileTransfer+AutoDownload.swift
