+++ Core/Barcelona/Attachments/IMFileTransfer+Bridging.swift
