+++ Core/Barcelona/Attachments/IMFileTransfer+StateBound.swift
