--- Core/Barcelona/Extensions/IMFileTransfer/IMFileTransfer+Creatable.swift
@@ -1,10 +1,0 @@
-//
-//  IMFileTransfer+Creatable.swift
-//  CoreBarcelona
-//
-//  Created by Eric Rabil on 9/11/20.
-//  Copyright © 2020 Eric Rabil. All rights reserved.
-//
-
-import Foundation
-import IMCore
